feat: update loading indicator for new pin designs
